web/api: Bind clients to access context
