+++ apps/web/src/lib/api.ts
@@ -1,495 +1,31 @@
-export type {
-  AgentInfo,
-  SmartTag,
-  FileActivityKind,
-  CostSource,
-  SessionHead,
-  SessionFileActivity,
-  FileActivityResult,
-  ProjectIdentityKind,
-  ProjectIdentity,
-  ProjectIdentityRef,
-  MessageTokens,
-  ToolPartStatus,
-  ToolPartState,
-  TextPart,
-  ReasoningPart,
-  PlanPart,
-  ImageDataPart,
-  ImageUrlPart,
-  ImagePart,
-  ToolPart,
-  MessagePart,
-  Message,
-  SessionDetail,
-  SessionListPage,
-  SessionReference,
-  ScanStatusEvent,
-  BackfillStatus,
-  AgentScanStatus,
-  DashboardAgentStat,
-  DashboardDailyBucket,
-  ModelDistributionEntry,
-  ModelCostEntry,
-  DashboardTotals,
-  DashboardPreviousTotals,
-  DashboardProjectStat,
-  DashboardRecentSession,
-  DashboardData,
-  AppConfig,
-  SearchResult,
-  SessionsUpdatedEvent,
-  ApiProjectGroup,
-  ApiProjectPage,
-  ApiProjectSummary,
-  ApiProjectAgentStat,
-  BookmarkRecord,
-  BookmarkView,
-} from "@codesesh/core/contract";
+export * from "./api-contract";
+export { ApiRequestError } from "./api-client";
 
-export { SMART_TAGS } from "@codesesh/core/contract";
+import { createApiClient } from "./api-client";
+import { createClientTelemetry } from "./client-telemetry";
+import { resolveRemoteAccess } from "./remote-access";
+import { createSessionUpdateSubscriber } from "./session-updates";
 
-import { sessionRoutePath } from "@codesesh/core/contract";
-import type {
-  AgentInfo,
-  ApiProjectGroup,
-  ApiProjectPage,
-  AppConfig,
-  BookmarkRecord,
-  BookmarkView,
-  DashboardData,
-  FileActivityKind,
-  ProjectIdentityKind,
-  ProjectIdentityRef,
-  ScanStatusEvent,
-  SearchResult,
-  SessionDetail,
-  SessionListPage,
-  SessionReference,
-  SessionHead,
-  SessionsUpdatedEvent,
-  SmartTag,
-} from "@codesesh/core/contract";
+const remoteAccess = resolveRemoteAccess();
+export const apiClient = createApiClient(remoteAccess);
 
-const REMOTE_ACCESS_QUERY_PARAM = "access_token";
-const REMOTE_ACCESS_STORAGE_KEY = "codesesh:remote-access-token";
-let remoteAccessToken: string | null = null;
+export const {
+  fetchConfig,
+  fetchScanStatus,
+  fetchAgents,
+  fetchProjects,
+  fetchProject,
+  fetchSessions,
+  fetchSessionData,
+  fetchDashboard,
+  fetchSearchResults,
+  fetchBookmarks,
+  upsertBookmark,
+  importBookmarks,
+  deleteBookmark,
+  upsertSessionAlias,
+  deleteSessionAlias,
+} = apiClient;
 
-export function initializeRemoteAccess(): void {
-  if (typeof window === "undefined") return;
-
-  const url = new URL(window.location.href);
-  const urlToken = url.searchParams.get(REMOTE_ACCESS_QUERY_PARAM);
-  if (urlToken) {
-    remoteAccessToken = urlToken;
-    try {
-      window.sessionStorage.setItem(REMOTE_ACCESS_STORAGE_KEY, urlToken);
-    } catch {}
-    url.searchParams.delete(REMOTE_ACCESS_QUERY_PARAM);
-    window.history.replaceState(
-      window.history.state,
-      "",
-      `${url.pathname}${url.search}${url.hash}`,
-    );
-    return;
-  }
-
-  try {
-    remoteAccessToken = window.sessionStorage.getItem(REMOTE_ACCESS_STORAGE_KEY);
-  } catch {
-    remoteAccessToken = null;
-  }
-}
-
-function withRemoteAccess(init: RequestInit = {}): RequestInit {
-  if (!remoteAccessToken) return init;
-  const headers = new Headers(init.headers);
-  headers.set("Authorization", `Bearer ${remoteAccessToken}`);
-  return { ...init, headers };
-}
-
-function remoteAccessUrl(path: string): string {
-  if (!remoteAccessToken) return path;
-  const url = new URL(path, window.location.origin);
-  url.searchParams.set(REMOTE_ACCESS_QUERY_PARAM, remoteAccessToken);
-  return `${url.pathname}${url.search}`;
-}
-
-export interface SearchRequestOptions {
-  agent?: string;
-  projectKind?: ProjectIdentityKind;
-  projectKey?: string;
-  tag?: SmartTag;
-  tool?: string;
-  fileKind?: FileActivityKind;
-  costMin?: number;
-  costMax?: number;
-  from?: number;
-  to?: number;
-}
-
-export interface FetchOptions {
-  signal?: AbortSignal;
-}
-
-export interface SessionDetailFetchOptions extends FetchOptions {
-  messageCursor?: string;
-}
-
-export interface ProjectPageOptions extends FetchOptions {
-  cursor?: string;
-  project?: ProjectIdentityRef;
-}
-
-interface SessionFetchProgress {
-  onFirstPage?: (sessions: SessionHead[]) => void;
-}
-
-const SESSION_PAGE_SIZE = 250;
-const SESSION_STALE_RETRY_LIMIT = 2;
-
-/**
- * Which slice of the snapshot a dashboard request covers. Project and agent are
- * independent and either may be absent: no project means every project, no agent
- * means every agent. A project dashboard is the global one with `project` set.
- */
-export interface DashboardFilters {
-  project?: { kind: ProjectIdentityKind; key: string };
-  agent?: string;
-}
-
-async function fetchJson<T>(path: string, init?: RequestInit): Promise<T> {
-  const res = await fetch(path, withRemoteAccess(init));
-  if (!res.ok) {
-    const method = init?.method ?? "GET";
-    throw new ApiRequestError(
-      `${method} ${path} failed: ${res.status} ${res.statusText}`,
-      res.status,
-    );
-  }
-  return res.json() as Promise<T>;
-}
-
-export class ApiRequestError extends Error {
-  constructor(
-    message: string,
-    readonly status: number,
-  ) {
-    super(message);
-  }
-}
-
-export async function fetchConfig(options?: FetchOptions): Promise<AppConfig> {
-  return fetchJson("/api/config", options);
-}
-
-export async function fetchScanStatus(): Promise<ScanStatusEvent> {
-  return fetchJson("/api/status");
-}
-
-function appendTimeWindow(params: URLSearchParams, window?: AppConfig["window"]): void {
-  if (window?.from != null) params.set("from", new Date(window.from).toISOString());
-  if (window?.to != null) params.set("to", new Date(window.to).toISOString());
-}
-
-export async function fetchAgents(
-  window?: AppConfig["window"],
-  options?: FetchOptions,
-): Promise<AgentInfo[]> {
-  const params = new URLSearchParams();
-  appendTimeWindow(params, window);
-  return fetchJson(`/api/agents?${params}`, options);
-}
-
-export async function fetchProjects(
-  window?: AppConfig["window"],
-  options?: ProjectPageOptions,
-): Promise<ApiProjectPage> {
-  const params = new URLSearchParams();
-  appendTimeWindow(params, window);
-  params.set("limit", options?.project ? "1" : "100");
-  if (options?.cursor) params.set("cursor", options.cursor);
-  if (options?.project) {
-    params.set("projectKind", options.project.kind);
-    params.set("projectKey", options.project.key);
-  }
-  return fetchJson(
-    `/api/projects?${params}`,
-    options?.signal ? { signal: options.signal } : undefined,
-  );
-}
-
-export async function fetchProject(
-  window: AppConfig["window"],
-  project: ProjectIdentityRef,
-  options?: FetchOptions,
-): Promise<ApiProjectGroup | null> {
-  const page = await fetchProjects(window, { ...options, project });
-  return page.projects[0] ?? null;
-}
-
-export async function fetchSessions(
-  options: {
-    agent?: string;
-    projectKind?: ProjectIdentityKind;
-    projectKey?: string;
-    from?: number;
-    to?: number;
-  } = {},
-  fetchOptions?: FetchOptions,
-  progress?: SessionFetchProgress,
-): Promise<{ sessions: SessionHead[] }> {
-  const baseParams = new URLSearchParams();
-  if (options.agent) baseParams.set("agent", options.agent);
-  if (options.projectKind) baseParams.set("projectKind", options.projectKind);
-  if (options.projectKey) baseParams.set("projectKey", options.projectKey);
-  appendTimeWindow(baseParams, options);
-
-  let staleRetries = 0;
-  while (true) {
-    try {
-      const sessions: SessionHead[] = [];
-      let cursor: string | undefined;
-      do {
-        const params = new URLSearchParams(baseParams);
-        params.set("limit", String(SESSION_PAGE_SIZE));
-        if (cursor) params.set("cursor", cursor);
-        const page = await fetchJson<SessionListPage>(`/api/sessions?${params}`, fetchOptions);
-        sessions.push(...page.sessions);
-        if (!cursor) progress?.onFirstPage?.([...sessions]);
-        cursor = page.nextCursor;
-      } while (cursor);
-      return { sessions };
-    } catch (error) {
-      if (
-        !(error instanceof ApiRequestError) ||
-        error.status !== 409 ||
-        staleRetries >= SESSION_STALE_RETRY_LIMIT
-      ) {
-        throw error;
-      }
-      staleRetries += 1;
-    }
-  }
-}
-
-export async function fetchSessionData(
-  agent: string,
-  sessionId: string,
-  options?: SessionDetailFetchOptions,
-): Promise<SessionDetail> {
-  const path = `/api/sessions${sessionRoutePath({ agentName: agent, sessionId })}`;
-  const fetchOptions: FetchOptions | undefined = options ? { signal: options.signal } : undefined;
-  if (!options?.messageCursor) return fetchJson(path, fetchOptions);
-
-  const params = new URLSearchParams({ messageCursor: options.messageCursor });
-  return fetchJson(`${path}?${params}`, fetchOptions);
-}
-
-export async function fetchDashboard(
-  window: AppConfig["window"] | undefined,
-  filters: DashboardFilters,
-  options?: FetchOptions,
-): Promise<DashboardData> {
-  const params = new URLSearchParams();
-  if (window?.days === 0) {
-    params.set("days", "0");
-  } else if (window?.from != null) {
-    appendTimeWindow(params, window);
-  } else {
-    if (window?.to != null) params.set("to", String(window.to));
-    if (window?.days != null) params.set("days", String(window.days));
-  }
-  if (filters.project) {
-    params.set("projectKind", filters.project.kind);
-    params.set("projectKey", filters.project.key);
-  }
-  if (filters.agent) params.set("agent", filters.agent);
-  const suffix = params.toString();
-  return fetchJson(suffix ? `/api/dashboard?${suffix}` : "/api/dashboard", options);
-}
-
-export async function fetchSearchResults(
-  query: string,
-  options: SearchRequestOptions = {},
-  fetchOptions?: FetchOptions,
-): Promise<{ results: SearchResult[] }> {
-  const params = new URLSearchParams();
-  params.set("q", query);
-  if (options.agent) params.set("agent", options.agent);
-  if (options.projectKind) params.set("projectKind", options.projectKind);
-  if (options.projectKey) params.set("projectKey", options.projectKey);
-  if (options.tag) params.set("tag", options.tag);
-  if (options.tool) params.set("tool", options.tool);
-  if (options.fileKind) params.set("fileActivity", options.fileKind);
-  if (options.costMin != null) params.set("costMin", String(options.costMin));
-  if (options.costMax != null) params.set("costMax", String(options.costMax));
-  appendTimeWindow(params, options);
-  return fetchJson(`/api/search?${params}`, fetchOptions);
-}
-
-export async function fetchBookmarks(
-  options?: FetchOptions,
-): Promise<{ bookmarks: BookmarkView[] }> {
-  return fetchJson("/api/bookmarks", options);
-}
-
-export async function upsertBookmark(
-  reference: SessionReference,
-): Promise<{ bookmark: BookmarkRecord }> {
-  return fetchJson("/api/bookmarks", {
-    method: "PUT",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ reference }),
-  });
-}
-
-export async function importBookmarks(
-  bookmarks: BookmarkRecord[],
-): Promise<{ bookmarks: BookmarkView[] }> {
-  return fetchJson("/api/bookmarks/import", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(bookmarks),
-  });
-}
-
-export async function deleteBookmark(reference: SessionReference): Promise<void> {
-  await fetchJson(`/api/bookmarks${sessionRoutePath(reference)}`, { method: "DELETE" });
-}
-
-export async function upsertSessionAlias(
-  agentKey: string,
-  sessionId: string,
-  alias: string,
-): Promise<void> {
-  await fetchJson(`/api/session-aliases${sessionRoutePath({ agentName: agentKey, sessionId })}`, {
-    method: "PUT",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ alias }),
-  });
-}
-
-export async function deleteSessionAlias(agentKey: string, sessionId: string): Promise<void> {
-  await fetchJson(`/api/session-aliases${sessionRoutePath({ agentName: agentKey, sessionId })}`, {
-    method: "DELETE",
-  });
-}
-
-export function logClientEvent(event: string, data: Record<string, unknown> = {}): void {
-  const body = JSON.stringify({ event, data });
-
-  try {
-    if (!remoteAccessToken && typeof navigator !== "undefined" && navigator.sendBeacon) {
-      const blob = new Blob([body], { type: "application/json" });
-      if (navigator.sendBeacon("/api/logs", blob)) return;
-    }
-  } catch {}
-
-  void fetch(
-    "/api/logs",
-    withRemoteAccess({
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body,
-      keepalive: true,
-    }),
-  ).catch(() => {});
-}
-
-const INITIAL_RETRY_MS = 1_000;
-const MAX_RETRY_MS = 30_000;
-const DISCONNECT_NOTICE_DELAY_MS = 5_000;
-
-function jitter(delayMs: number): number {
-  const spread = delayMs * 0.2;
-  return delayMs + (Math.random() * 2 - 1) * spread;
-}
-
-export function subscribeSessionUpdates(
-  onUpdate: (event: SessionsUpdatedEvent) => void,
-  onScanStatus?: (event: ScanStatusEvent) => void,
-  onReconnect?: () => void,
-  onDisconnect?: () => void,
-): () => void {
-  let disposed = false;
-  let retryTimer: ReturnType<typeof setTimeout> | undefined;
-  let disconnectTimer: ReturnType<typeof setTimeout> | undefined;
-  let retryDelayMs = INITIAL_RETRY_MS;
-  let hasConnectedOnce = false;
-  let disconnectNotified = false;
-  let currentSource: EventSource | undefined;
-
-  const clearDisconnectTimer = () => {
-    if (disconnectTimer === undefined) return;
-    clearTimeout(disconnectTimer);
-    disconnectTimer = undefined;
-  };
-
-  const notifyDisconnect = () => {
-    if (disconnectNotified) return;
-    disconnectNotified = true;
-    onDisconnect?.();
-  };
-
-  const connect = () => {
-    const source = new EventSource(remoteAccessUrl("/api/events"));
-    currentSource = source;
-
-    source.addEventListener("sessions-updated", (event) => {
-      try {
-        onUpdate(JSON.parse(event.data) as SessionsUpdatedEvent);
-      } catch (error) {
-        console.error("Failed to parse session update event:", error);
-      }
-    });
-
-    source.addEventListener("scan-status", (event) => {
-      if (!onScanStatus) return;
-      try {
-        onScanStatus(JSON.parse(event.data) as ScanStatusEvent);
-      } catch (error) {
-        console.error("Failed to parse scan status event:", error);
-      }
-    });
-
-    source.onopen = () => {
-      clearDisconnectTimer();
-      retryDelayMs = INITIAL_RETRY_MS;
-      const recoveredFromDisconnect = disconnectNotified;
-      disconnectNotified = false;
-      if (hasConnectedOnce || recoveredFromDisconnect) {
-        onReconnect?.();
-      }
-      hasConnectedOnce = true;
-    };
-
-    source.onerror = () => {
-      if (source.readyState !== EventSource.CLOSED) {
-        disconnectTimer ??= setTimeout(() => {
-          disconnectTimer = undefined;
-          if (!disposed && currentSource === source) notifyDisconnect();
-        }, DISCONNECT_NOTICE_DELAY_MS);
-        return;
-      }
-      clearDisconnectTimer();
-      source.close();
-      if (disposed) return;
-      notifyDisconnect();
-      const delay = jitter(retryDelayMs);
-      retryDelayMs = Math.min(retryDelayMs * 2, MAX_RETRY_MS);
-      retryTimer = setTimeout(connect, delay);
-    };
-  };
-
-  connect();
-
-  return () => {
-    disposed = true;
-    if (retryTimer !== undefined) {
-      clearTimeout(retryTimer);
-    }
-    clearDisconnectTimer();
-    currentSource?.close();
-  };
-}
+export const { logClientEvent } = createClientTelemetry(remoteAccess);
+export const subscribeSessionUpdates = createSessionUpdateSubscriber(remoteAccess);
