+++ apps/web/src/main.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { RenderProfiler } from "./components/RenderProfiler.tsx";
-import { initializeRemoteAccess } from "./lib/api.ts";
 import { queryClient } from "./lib/query-client.ts";
 import { AppRouter } from "./router.tsx";
 import { ScanStatusProvider } from "./hooks/useScanStatus.tsx";
@@ -10,8 +9,6 @@ import "@fontsource-variable/space-grotesk";
 import "@fontsource-variable/jetbrains-mono";
 import "@fontsource-variable/ibm-plex-sans";
 import "./index.css";
-
-initializeRemoteAccess();
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
